🎉Release 1.0.0 🎉
This release is dedicated to the team at VRChat
responsible for shipping the UNU update, including,
but not limited to, Danly, MomoTheMonster, and the
VRC QA team (including Ruubick, PhaseDragon, Jordo
and others). On this rock we shall build our church.

In addition, thank you to the VRC Prefabs voice call
gang for being very dear friends of mine, to
Senpie and PairOfSocks, and everyone else who
played the heck out of the ht8b prefab these past
five months.