New metal-frame table variant! (Thanks Ivylistar!)
This massive commit introduces a new table to the repo!

It's sleek, metal, and comes with a SICK glass face which exposes its intricate under-frame, which allows pocketed balls to roll down and collect in the center underneath.
(NOTE: THIS DOESN'T WORK YET AND REQUIRES A CODE CHANGE)

I'm also including Silent's Filamented shader, as it's just plain better than Standard in many use-cases (including ours) and makes our new table look 👌

There's still some work to be done with little touches but this commit *should* be basically functional and production-ready at this point. (aside from the balls not staying under the table's frame and the glass emission color not changing)