Initial work getting GPU Instancing working with ghost indicators
Since there's multiple of them visible at any given time even with *one* table in the world, I figured it'd be valuable to have them be instanced wherever possible.

I've never actually modified a shader to work with instancing before so I'm not sure if this is *ideal*. It DOES seem to work with randomly placed spheres, but doesn't appear to work with the actual icospheres on the prefab (yet). Will likely figure it out later..