Fixes to support Ivy's cool glass-topped table.
To support Ivy's glass-topped table properly, the
prefabs were messed with a little to support further
nesting. Maybe this bites us in the bum in the future,
but nesting prefabs are more modular and are
generally the go-to solution for complex prefabs.

In addition, code changes were introduced to improve
on the original RackBalls code. This code is more flexible,
more declarative and also allows for more table modding
by the end user.