Clamped ball velocity when sunk.
Clamped the max velocity of balls when sunk to -1, 1 on both X and Z.

This means balls flying off the table is quite unlikely now.

Also moved the spawn of sunk balls a little inside the wood table to help prevent clipping when all 16 balls are in a sunk state.