Hotfix: Removed an unused bool array from PoolStateManager
Arrays are objects, not primitives, ergo their default value is
null. Udon, in its infinite wisdom, has no guard for what happens
when someone tries to sync a null array. I'm unsure if this is
because the devs forgot the default for an array is null or not,
but regardless it is a nasty bug.

This array is to be used for the slow-boil approach of getting rid
of ballPocketedState, but it's unused right now and is safe to remove.