Improve dependency management for PoolStateManager for easier prefabbing.
Nested prefabs being configured by their parent prefab can be modified,
but this can get sticky with Udon. Assume that the base object for the
core script is the grandparent of the core script object, and assume
the menu is in one of the grandparent's descendents. If either of these
assumptions fail, terminate the table's code.

Also, clean up the options section at the start of the PoolStateManager
script.