Make respawning work off of local position/rotation
The cues used to use world position and rotation to figure out
where they go when they respawn. This is wrong - we should use
localPosition, which then is nice and consistent even if the
table has been moved for some reason.

This requires a VR test as it touches the traditionally flaky
respawn code.