Loggers are now off (and EditorOnly) by default
Most people hide/move the logger anyway, and it serves little purpose nowadays. If you want it in your world, you can easily untag it and show the gameobject. At some point in the future, we'll likely have an editor script which asks you questions at setup and makes all of this easy peasy.