Fix bad "first/second/third ball hit" var reset
Seems like a longer-term bug regarding scoring
was broken where the table would flash a
bad colour or score a point when you shouldn't
if you were not playing. I've made an attempt
to fix this.

I've also cleaned up the code a little and removed
that bad FixedUpdate call that seems to just
increase CPU usage for no benefit.