Kill off UdonChips 🗡️
This is a rarely-used Prefab to begin with and
the implementation is hacky. Proper support for
sidecars (such as in-world currency systems)
is forthcoming, but we won't be supporting
betting and gambling in the table in the
future; this would risk someone rigging up
a Creator Economy-fueled gambling ring
using our pool table, and I don't want
to endure the Kotaku article.

I asked Table about this and they replied
with a weird cat doing a thumbs up emoji