New Physics, Other Upgrades, 2022 update (#186)
* The New Physics System Backport

This work was on the 1.6 experimental repo for a while
and I'm backporting it back into the main repo, as it
seems to be broadly stable and it's a major feature
win for the formal 1.6 update when it happens.

* Some script cleanup

* Some prefab updates

* Add Kofi sponsor link

* WIP progress

* Kill off UdonChips 🗡️

This is a rarely-used Prefab to begin with and
the implementation is hacky. Proper support for
sidecars (such as in-world currency systems)
is forthcoming, but we won't be supporting
betting and gambling in the table in the
future; this would risk someone rigging up
a Creator Economy-fueled gambling ring
using our pool table, and I don't want
to endure the Kotaku article.

I asked Table about this and they replied
with a weird cat doing a thumbs up emoji

* Finalize example Sci-Fi prefab for validation

Enough here to be going along with! A number
of small and large changes, especially to the
desktop UI.

* It's slime time

* Kick it to 2022

* Correctly set up OscSlogger

* Complete Prefab migration; some code fixes

A lot of misc code fixes to support moving the four visually
distinct tables (regular, metal, cottonfox, sci-fi) to
1.6.0