Rapid-cue-switch crash bug mitigation (#201)
* Update BasePoolStateManager.cs

Sets previous frame cue tip position to current position on local turn start.  Prevents the ball from being launched halfway across the map  if the player manages to arm the cue with the cue tip already touching the cue ball.

* Update PoolStateManager.cs

Checks contents of Sqrt calculations to ensure they are >= 0. 
Prevents a VRchat crash in the event that a player manages to arm the cue with the tip far enough inside the ball to create an invalid state.